modify SmartContractAutofacModule
