+++ AElf.Kernel.Tests/ConfigureTestFramework.cs
@@ -44,7 +44,7 @@ namespace AElf.Kernel.Tests
             
             builder.RegisterModule(new LoggerModule());
             builder.RegisterModule(new DatabaseAutofacModule());
-            builder.RegisterModule(new MetadataModule());
+            builder.RegisterModule(new SmartContractAutofacModule());
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new StorageModule());
             builder.RegisterModule(new ServicesModule());
