+++ AElf.Miner.Tests/ConfigureTestFramework.cs
@@ -46,7 +46,7 @@ namespace AElf.Miner.Tests
             
             builder.RegisterModule(new LoggerModule());
             builder.RegisterModule(new DatabaseAutofacModule());
-            builder.RegisterModule(new MetadataModule());
+            builder.RegisterModule(new SmartContractAutofacModule());
             builder.RegisterModule(new StateDictatorModule());
             builder.RegisterModule(new StorageModule());
             builder.RegisterModule(new ServicesModule());
