+++ AElf.SmartContract/SmartContractAutofacModule.cs
@@ -1,4 +1,5 @@
-﻿using Autofac;
+﻿using AElf.SmartContract.Metadata;
+using Autofac;
 
 namespace AElf.SmartContract
 {
@@ -6,7 +7,13 @@ namespace AElf.SmartContract
     {
         protected override void Load(ContainerBuilder builder)
         {
+            var assembly1 = typeof(IStateDictator).Assembly;
+            builder.RegisterAssemblyTypes(assembly1).AsImplementedInterfaces();
             
+            var assembly2 = typeof(StateDictator).Assembly;
+            builder.RegisterAssemblyTypes(assembly2).AsImplementedInterfaces();
+            
+            builder.RegisterType(typeof(FunctionMetadataService)).As<IFunctionMetadataService>().SingleInstance();
         }
     }
 }
